# specs/api_contracts.yaml
openapi: "3.1.0"
info:
  title: ChatBridge v2 API
  version: "1.0.0"
  description: "K-12 AI chat platform with third-party app integration. Fastify backend with auto-generated OpenAPI (L-001)."
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /classrooms:
    post:
      summary: Create classroom
      tags: [Classrooms]
      operationId: createClassroom
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, grade_band, ai_config]
              properties:
                name: { type: string }
                grade_band: { type: string, enum: [k2, g35, g68, g912] }
                ai_config: { type: object }
      responses:
        "201": { description: "Classroom created with join code" }
      x-requirement: FR-CTRL-001
  /classrooms/{classroomId}/config:
    get:
      summary: Get classroom config (grade band, AI settings)
      tags: [Classrooms]
      operationId: getClassroomConfig
      responses:
        "200": { description: "Classroom config including grade_band and ai_config" }
      x-requirement: FR-CHAT-006

    patch:
      summary: Update classroom config
      tags: [Classrooms]
      operationId: updateClassroomConfig
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ai_config: { type: object }
                async_guidance: { type: string }
      responses:
        "200": { description: "Config updated" }
      x-requirement: FR-CTRL-004
...
  /classrooms/{classroomId}/apps/{appId}:
    patch:
      summary: Toggle app enabled/disabled for classroom
      tags: [Classrooms]
      operationId: toggleClassroomApp
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled: { type: boolean }
      responses:
        "200": { description: "App toggled" }
      x-requirement: FR-CTRL-002
  /classrooms/{classroomId}/students/{studentId}/whisper:
    post:
      summary: Send teacher whisper for specific student
      tags: [Classrooms]
      operationId: sendWhisper
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [guidance]
              properties:
                guidance: { type: string }
      responses:
        "200": { description: "Whisper stored for next AI response" }
      x-requirement: FR-CTRL-003
...
  /collaborative-sessions:
    post:
      summary: Create collaborative session
      tags: [Collaboration]
      operationId: createSession
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [app_id, interaction_model]
              properties:
                app_id: { type: string, format: uuid }
                interaction_model: { type: string, enum: [turn_based, concurrent] }
      responses:
        "201": { description: "Session created with code" }
      x-requirement: FR-COLLAB-001
  /collaborative-sessions/{code}/join:
    post:
      summary: Join collaborative session
      tags: [Collaboration]
      operationId: joinSession
      responses:
        "200": { description: "Joined with participant info" }
        "404": { description: "Session not found" }
        "409": { description: "Session full or closed" }
      x-requirement: FR-COLLAB-001

  # === Admin ===
...